Voxel changed from 16 bit to 32 bit + new lua functions

+++ src/typedefs.h
@@ -9,8 +9,8 @@ typedef unsigned int uint;
 // use for bytes arrays
 typedef uint8_t ubyte;
 
-typedef uint8_t blockid_t;
-typedef uint8_t blockstate_t;
+typedef uint16_t blockid_t;
+typedef uint16_t blockstate_t;
 typedef uint16_t light_t;
 
 #endif
